update pada controller byDaniel

+++ controllers/bookmarks.js
@@ -38,6 +38,24 @@ class Bookmark {
   }
   async getDetailBookmark(req, res, next) {
     try {
+      const token = req.headers.authorization.replace('Bearer ', '');
+      const currentUser = await user.findOne({
+          where: { token }
+      });
+
+      const currentBookmark = await bookmark.findOne({ where: { id: req.params.id } });
+      // if theres nothing here
+      if (currentBookmark === null) {
+          return res.status(404).json({ status: 500, message: 'Bookmark not found' });
+      }
+
+      if (currentUser.id != currentBookmark.id_user) {
+          return res.status(404).json({ errors: ['No access to this bookmark!'] });
+      }
+
+      if (currentUser == null) {
+          return res.status(404).json({ errors: ['No access to this bookmark!'] });
+      }
       let data = await bookmark.findOne({
         // find all data of bookmark table
         where: { id: req.params.id },
@@ -61,11 +79,6 @@ class Bookmark {
         ],
       });
 
-      // if theres nothing here
-      if (!data) {
-        return res.status(404).json({ errors: ["Bookmark is not found"] });
-      }
-
       // if success
       res.status(200).json({ data });
     } catch (error) {
@@ -75,6 +88,11 @@ class Bookmark {
 
   async createBookmark(req, res, next) {
     try {
+      const token = req.headers.authorization.replace('Bearer ', '');
+      const currentUser = await user.findOne({
+          where: { token }
+      });
+      req.body.id_user = currentUser.id;
       // create bookmark
       const newData = await bookmark.create(req.body);
 
@@ -111,17 +129,35 @@ class Bookmark {
   // update data
   async updateBookmark(req, res, next) {
     try {
+      const token = req.headers.authorization.replace('Bearer ', '');
+      const currentUser = await user.findOne({
+          where: { token }
+      });
+      const currentBookmark = await bookmark.findOne(
+          { 
+              where: {
+                  id: req.params.id
+              }
+          }
+      );
+
+      if (currentBookmark === null) {
+          return res.status(404).json({ status: 500, message: 'Bookmark not found' });
+      }
+
+      if (currentUser.id != currentBookmark.id_user) {
+          return res.status(404).json({ errors: ['No delete access to this bookmark!'] });
+      }
+
+      if (currentUser == null) {
+          return res.status(404).json({ errors: ['No delete access to this bookmark!'] });
+      }
       // bookmark table update
-      const updatedData = await bookmark.update(req.body, {
+      await bookmark.update(req.body, {
         where: {
           id: req.params.id,
         },
       });
-
-      // if no data updated
-      if (updatedData[0] === 0) {
-        return res.status(404).json({ errors: ["Bookmark not found"] });
-      }
 
       // find the updated bookmark
       const data = await bookmark.findOne({
@@ -158,13 +194,31 @@ class Bookmark {
   // delete data
   async deleteBookmark(req, res, next) {
     try {
-      // delete data
-      let data = await bookmark.destroy({ where: { id: req.params.id } });
+      const token = req.headers.authorization.replace('Bearer ', '');
+      const currentUser = await user.findOne({
+          where: { token }
+      });
+      const currentBookmark = await bookmark.findOne(
+          { 
+              where: {
+                  id: req.params.id
+              }
+          }
+      );
 
-      // If data deleted is null
-      if (!data) {
-        return res.status(404).json({ errors: ["Bookmark not found"] });
+      if (currentBookmark === null) {
+          return res.status(404).json({ status: 500, message: 'Bookmark not found' });
       }
+
+      if (currentUser.id != currentBookmark.id_user) {
+          return res.status(404).json({ errors: ['No delete access to this bookmark!'] });
+      }
+
+      if (currentUser == null) {
+          return res.status(404).json({ errors: ['No delete access to this bookmark!'] });
+      }
+      // delete data
+      await bookmark.destroy({ where: { id: req.params.id } });
 
       // if success
       res.status(200).json({ message: "Success delete bookmark" });
