+++ controllers/users.js
@@ -17,7 +17,7 @@ class User {
             if (data == null) {
                 let { firstName, lastName, email, password, avatar } = req.body;
                 if (!/^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/.test(email)) return res.status(400).json(["Masukkan email dengan benar!"]);
-                if (password.length < 10 || password.length > 20) return res.status(400).json([`Password tidak boleh ${password.length < 10 ? "kurang" : "lebih"} dari ${password.length < 10 ? "10" : "20"} karakter!`]);
+                if (!/^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[!@#\$%\^&\*\.\,])(?=.{10,20})/.test(password)) return res.status(400).json([`Masukkan password dengan benar!`]);
                 password = encrypt(password)
                 const newData = await user.create({
                     firstName, lastName, password, email, avatar,
@@ -78,27 +78,64 @@ class User {
     // updating user
     async updateUser(req, res, next) {
         try {
-            const updateUser = await user.update({
-                where: { 
-                    id: req.params.id
-                },
-                attributes: { exclude: ["createdAt", "deletedAt", "password"]}
-            })
-            if(updateUser[0] === 0) {
-                return res.status(404).json({ errors: ['User not found!'] });
+            const token = req.headers.authorization.replace('Bearer ', '');
+            const currentUser = await user.findOne({
+                where: { token }
+            });
+            
+            if (currentUser.id != req.params.id) {
+                return res.status(404).json({ errors: ['No edit access to this user!'] });
             }
-            const data = await user.findOne({where: {id: req.params.id}});
+
+            if (currentUser == null) {
+                return res.status(404).json({ errors: ['No edit access to this user!'] });
+            }
+
+            else {
+                let { firstName, lastName, email, password, avatar } = req.body;
+                if (!/^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/.test(email)) return res.status(400).json(["Masukkan email dengan benar!"]);
+                if (!/^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[!@#\$%\^&\*\.\,])(?=.{10,20})/.test(password)) return res.status(400).json([`Masukkan password dengan benar!`]);
+                password = encrypt(password)
+                await user.update(
+                    { firstName, lastName, email, password, avatar },
+                    { where: { id: currentUser.id }
+                });
+            }
+            // const updateUser = await user.update({
+            //     where: { 
+            //         id: req.params.id
+            //     },
+            //     attributes: { exclude: ["createdAt", "deletedAt", "password"]}
+            // })
+            // if(updateUser[0] === 0) {
+            //     return res.status(404).json({ errors: ['User not found!'] });
+            // }
+            const data = await user.findOne({where: { id: currentUser.id }});
 
             res.status(201).json({ data });
         } catch (error) {
+            console.error(error)
             res.status(500).json(['Error updating USER']);
         }
     }
     // deleting user
     async deleteUser(req, res, next) {
         try {
+            const token = req.headers.authorization.replace('Bearer ', '');
+            const currentUser = await user.findOne({
+                where: { token }
+            });
+            
+            if (currentUser.id != req.params.id) {
+                return res.status(404).json({ errors: ['No delete access to this user!'] });
+            }
+
+            if (currentUser == null) {
+                return res.status(404).json({ errors: ['No delete access to this user!'] });
+            }
+
             const data = await user.destroy({
-                where: { id: req.params.id },
+                where: { id: currentUser.id },
                 attributes: { exclude: ["createdAt", "updatedAt", "password"] },
             })
 
@@ -108,7 +145,7 @@ class User {
 
             res.status(200).json({ messages: ['User account has successfully been deleted!'] });
         } catch (error) {
-            res.status(500).json(['Error deleting USER']);
+            res.status(500).json(['Error updating USER']);
         }
     }
     // =============================================================================================================
@@ -125,7 +162,7 @@ class User {
         // cek apa password memang dimiliki oleh email yang diinput
         let validPass = decrypt(req.body.password, data.password);
         if(!validPass) {
-            return res.status(404).json({ errors: ['Kata Sandi yang dimasukkan salah']})
+            return res.status(404).json({ errors: ['Kata sandi yang dimasukkan salah']})
         }
 
         const {
@@ -137,6 +174,7 @@ class User {
         }
 
         const token = createToken(tmpToken);
+        await sequelize.query(`UPDATE users SET token='${token}' WHERE id=${data.id}`);
         return res.status(200).json({ statusCode: 200, message: "Login success!", token });
     }
